backend/requirements.txt
typer[all]>=0.9.0
rich>=13.0.0
requests>=2.31.0
PyYAML>=6.0.0
